pyproject.toml
[project]
name = "movie-recommender-project"
version = "0.1.0"
description = "Recommends movies based on input text"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.3",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "pinecone[grpc]>=5.4.2",
    "python-dotenv>=1.0.1",
]
